Refactor file directory trees
